# openapi/kairo-daemon.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kairo-daemon",
    "description": "Read-only HTTP+JSON API over a local Kairo store.\n\nAll non-streaming responses are wrapped in an envelope:\n`{ \"ok\": true, \"schema\": \"kairo.api.result.v1\", \"result\": <T> }`.\nErrors use the same envelope with `ok: false` and an `error` body\ncarrying a stable `code` and a human `message` (see `specs/API.md` §8).\nThe `body = T` declarations on each path describe the *inner* result\ntype; clients are expected to unwrap the envelope before consuming.\n\nThe blob endpoint (`GET /api/v1/blobs/{id}`) bypasses the envelope\nentirely and streams `application/octet-stream` bytes.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/capabilities/for-object/{object}": {
      "get": {
        "tags": [
          "capabilities"
        ],
        "summary": "`GET /api/v1/capabilities/for-object/{object_id}`",
        "operationId": "listCapabilitiesForObject",
        "parameters": [
          {
            "name": "object",
            "in": "path",
            "description": "Object id (kairo:object:...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Capability head summaries scoped to this object, hydrated with each head's scope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CapabilityHeadDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed object id"
          }
        }
      }
    },
    "/api/v1/capabilities/{grantor}": {
      "get": {
        "tags": [
          "capabilities"
        ],
        "operationId": "listCapabilitiesFromGrantor",
        "parameters": [
          {
            "name": "grantor",
            "in": "path",
            "description": "Grantor actor id (kairo:actor:...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Capability head summaries: one per (grantee, scope) chain leaf",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CapabilityHeadDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed grantor id"
          }
        }
      }
    },
...
    "/api/v1/trust/about/{of}": {
      "get": {
        "tags": [
          "trust"
        ],
        "summary": "`GET /api/v1/trust/about/{of}`",
        "operationId": "listTrustAbout",
        "parameters": [
          {
            "name": "of",
            "in": "path",
            "description": "Actor whose incoming opinions to list (kairo:actor:...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chain-leaf trust heads expressed about the actor (one per by_actor)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrustHeadDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed actor id"
          }
        }
      }
    },
    "/api/v1/trust/{by}/{of}": {
      "get": {
        "tags": [
          "trust"
        ],
        "operationId": "getTrust",
        "parameters": [
          {
            "name": "by",
            "in": "path",
            "description": "Actor whose opinion to read (kairo:actor:...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "of",
            "in": "path",
            "description": "Actor whose trust state to look up",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest ActorTrust statement (grant, block, or withdrawal)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorTrustStatementJson"
                }
              }
            }
          },
          "400": {
            "description": "Malformed actor id"
          },
          "404": {
            "description": "No trust opinion recorded; treat as `unknown`"
          }
        }
      }
    },
...
    "/api/v1/version-tags/{object}": {
      "get": {
        "tags": [
          "version-tags"
        ],
        "summary": "`GET /api/v1/version-tags/{object_id}`",
        "operationId": "listVersionTags",
        "parameters": [
          {
            "name": "object",
            "in": "path",
            "description": "Object id (kairo:object:...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Version-tag head summaries: one per (actor, version) chain leaf",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersionTagHeadDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed object id"
          }
        }
      }
    },
    "/api/v1/version-tags/{object}/{version}": {
      "get": {
        "tags": [
          "version-tags"
        ],
        "operationId": "getLatestVersionTag",
        "parameters": [
          {
            "name": "object",
            "in": "path",
            "description": "Object id (kairo:object:...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "Version label (e.g., \"v1.0.0\")",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "description": "Actor whose tag chain to resolve; defaults to the object's `created_by`",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest ObjectVersionTag statement honoring cross-actor supersedes flips",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectVersionTagStatementJson"
                }
              }
            }
          },
          "400": {
            "description": "Malformed object id, version, or actor query"
          },
          "404": {
            "description": "Version tag head not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ActorTrustBodyJson": {
        "type": "object",
        "required": [
          "trusted_actor"
        ],
        "properties": {
          "decision": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "supersedes": {
            "type": [
              "string",
              "null"
            ]
          },
          "trusted_actor": {
            "type": "string"
          }
        }
      },
      "ActorTrustStatementJson": {
        "type": "object",
        "required": [
          "type",
          "version",
          "actor",
          "subject",
          "created_at",
          "body",
          "signature"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/ActorTrustBodyJson"
          },
          "created_at": {
            "type": "string"
          },
          "signature": {
            "$ref": "#/components/schemas/SignatureJson"
          },
          "subject": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "CapabilityHeadDto": {
        "type": "object",
        "description": "Capability head summary returned by\n`GET /api/v1/capabilities/{grantor}` and\n`GET /api/v1/capabilities/for-object/{object}`. One entry\nper `(grantee, scope)` chain leaf — mirrors the shape\nproduced by `kairo capability list --grantor` in direct\nmode.",
        "required": [
          "grantor",
          "grantee",
          "scope",
          "statement_id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "grantee": {
            "type": "string"
          },
          "grantor": {
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/CapabilityScopeJson"
          },
          "statement_id": {
            "type": "string"
          }
        }
      },
...
      "CapabilityScopeJson": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "object"
            ],
            "properties": {
              "object": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "actor"
            ],
            "properties": {
              "actor": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "ObjectVersionTagBodyJson": {
        "type": "object",
        "required": [
          "object",
          "version"
        ],
        "properties": {
          "object": {
            "type": "string"
          },
          "supersedes": {
            "type": [
              "string",
              "null"
            ]
          },
          "target": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": "string"
          }
        }
      },
      "ObjectVersionTagStatementJson": {
        "type": "object",
        "required": [
          "type",
          "version",
          "actor",
          "subject",
          "created_at",
          "body",
          "signature"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/ObjectVersionTagBodyJson"
          },
          "created_at": {
            "type": "string"
          },
          "signature": {
            "$ref": "#/components/schemas/SignatureJson"
          },
          "subject": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "SignatureJson": {
        "type": "object",
        "required": [
          "actor",
          "key_id",
          "algorithm",
          "bytes"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "algorithm": {
            "type": "string"
          },
          "bytes": {
            "type": "string"
          },
          "key_id": {
            "type": "string"
          }
        }
      },
...
      "TrustHeadDto": {
        "type": "object",
        "description": "Trust head summary returned by\n`GET /api/v1/trust/about/{actor}`. One entry per `by_actor`\nchain leaf — callers who want the signed `ActorTrust` body\nfollow up with `GET /api/v1/trust/{by}/{of}` or\n`GET /api/v1/statements/{statement_id}`.\n\n`decision` is `\"trusted\"`, `\"untrusted\"`, or `null` (the\nlatter encodes a withdrawal — first-person \"no opinion\",\ndistinct from \"never expressed an opinion\").",
        "required": [
          "by_actor",
          "trusted_actor",
          "statement_id",
          "created_at"
        ],
        "properties": {
          "by_actor": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "description": "RFC 3339 UTC seconds."
          },
          "decision": {
            "type": [
              "string",
              "null"
            ]
          },
          "statement_id": {
            "type": "string"
          },
          "trusted_actor": {
            "type": "string"
          }
        }
      },
...
      "VersionTagHeadDto": {
        "type": "object",
        "description": "Version-tag head summary returned by\n`GET /api/v1/version-tags/{object}`. Light shape, one entry\nper `(actor, version)` chain leaf — callers who want the\nsigned tag follow up with\n`GET /api/v1/version-tags/{object}/{version}` or\n`GET /api/v1/statements/{statement_id}`.",
        "required": [
          "actor",
          "object",
          "version",
          "statement_id",
          "created_at"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "description": "RFC 3339 UTC seconds."
          },
          "object": {
            "type": "string"
          },
          "statement_id": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      }
    }
  },
...
}